requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-chroma>=0.2.0
langgraph>=0.4.0
chromadb>=1.0.0
google-genai>=1.70.0
pymupdf>=1.25.0
Pillow>=11.0.0
python-dotenv>=1.1.0
streamlit>=1.45.0
pytest>=8.0.0
matplotlib>=3.9.0
streamlit-mermaid>=0.4.0
